Update DTOs; Add WarehousesService

+++ src/main/resources/db/changelog/v3/v3-2-change-warehouses-table.sql
@@ -1,4 +1,4 @@
 --liquibase formatted sql
 --changeset ybryak:3-2-change-warehouses-table
 ALTER TABLE warehouses
-CHANGE owner owner_id INT;
+CHANGE owner owner_id LONG;
